+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.2]
+
 ### Added:
 - Logging of directory listings
 - Buffering
@@ -104,7 +106,8 @@ Paths that previously returned 404 may no longer.
 
 
 
-[unreleased]: https://github.com/Jacoblightning/SimpleWebServer-RS/compare/v2.2.1...main
+[unreleased]: https://github.com/Jacoblightning/SimpleWebServer-RS/compare/v2.2.2...main
+[2.3.0]: https://github.com/Jacoblightning/SimpleWebServer-RS/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/Jacoblightning/SimpleWebServer-RS/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/Jacoblightning/SimpleWebServer-RS/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/Jacoblightning/SimpleWebServer-RS/compare/v2.0.0...v2.1.0